You now mixin a ExecutionContexts trait rather than use the implicit import.
This gives us a bit more control over the execution context
