+++ core-navigation/app/conf/context.scala
@@ -6,6 +6,7 @@ import com.gu.management.{ PropertiesPage, StatusPage, ManifestPage, Switchable,
 import com.gu.management.play.{ Management => GuManagement }
 
 import com.gu.management.logback.LogbackLevelPage
+import contentapi.ContentApiMetrics
 
 object Switches {
   val all: Seq[Switchable] = CommonSwitches.all
