+++ football/app/controllers/FrontScoresController.scala
@@ -4,9 +4,9 @@ import common._
 import feed.Competitions
 import play.api.mvc.{ Action, Controller }
 import model._
-import play.api.libs.concurrent.Execution.Implicits._
 
-object FrontScoresController extends Controller with implicits.Football with Logging {
+
+object FrontScoresController extends Controller with implicits.Football with Logging with ExecutionContexts {
 
   /*
    * Finds the first competition with matches on today in a list of competition ids passed in
