+++ gallery/app/controllers/GalleryController.scala
@@ -5,7 +5,7 @@ import common._
 import conf._
 import model._
 import play.api.mvc.{ RequestHeader, Controller, Action }
-import play.api.libs.concurrent.Execution.Implicits._
+
 import concurrent.Future
 
 case class GalleryPage(
@@ -14,7 +14,7 @@ case class GalleryPage(
   index: Int,
   trail: Boolean)
 
-object GalleryController extends Controller with Logging {
+object GalleryController extends Controller with Logging with ExecutionContexts {
 
   def render(path: String) = Action { implicit request =>
