+++ style-guide/app/controllers/StyleGuideController.scala
@@ -5,7 +5,7 @@ import model._
 import conf._
 import com.gu.openplatform.contentapi.model.ItemResponse
 import play.api.mvc.{ Content => _, _ }
-import play.api.libs.concurrent.Execution.Implicits._
+
 import concurrent.Future
 
 
@@ -13,7 +13,7 @@ case class ArticlePage(article: Article, storyPackage: List[Trail])
 
 case class ZoneColour(className: String, selectors: List[String], zoneName: String, hexCode: String)
 
-object StyleGuideController extends Controller with Logging {
+object StyleGuideController extends Controller with Logging with ExecutionContexts {
 
   def renderIndex = Action { implicit request =>
     val page = Page(canonicalUrl = None, "style-guide", "style-guide", "Style guide: home", "GFE:Style-guide")
