+++ tag/app/conf/context.scala
@@ -5,6 +5,7 @@ import common._
 import com.gu.management._
 import com.gu.management.play._
 import logback.LogbackLevelPage
+import contentapi.ContentApiMetrics
 
 object Switches {
   val all: Seq[Switchable] = CommonSwitches.all // ++ new DefaultSwitch("name", "Description Text")
